requirements.txt
smolagents
streamlit
python-dotenv
litellm
tensorflow
yfinance
pydantic